+++ astropy/io/fits/tests/test_diff.py
@@ -576,17 +576,3 @@ class TestDiff(FitsTestCase):
         assert diff.diff_values[1][0] == ('colb', 1)
         assert np.isnan(diff.diff_values[1][1][0])
         assert diff.diff_values[1][1][1] == 2.0
-
-        # What about in Headers?
-
-        h = fits.Header([('A', 1), ('B', np.nan)])
-        h2 = h.copy()
-
-        assert HeaderDiff(h, h2).identical
-
-        h2['B'] = 1.0
-
-        diff = HeaderDiff(h, h2)
-        assert not diff.identical
-        assert np.isnan(diff.diff_keyword_values['B'][0][0])
-        assert diff.diff_keyword_values['B'][0][1] == 1.0